pyproject.toml
[project]
name = "glowing-octo-robot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "open_clip_torch>=2.24",
    "numpy>=2.3.1",
    "pillow>=11.2.1",
    "pyside6>=6.9.1",
    "torch>=2.7.1",
    "torchvision>=0.22.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/image_partition"]

[tool.uv.sources]
torch = [
  { index = "pytorch-cu126", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]
torchvision = [
  { index = "pytorch-cu126", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]

[[tool.uv.index]]
name = "pytorch-cu126"
url = "https://download.pytorch.org/whl/cu126"
explicit = true

[tool.uv]
dev-dependencies = [
  "black",
  "ruff",
  "pytest",
  "pytest-qt",
  "mypy",
]
